minor cleanup add .gitignore
- update deploy.sh

+++ deploy.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+# .git/refs/heads/master >deploy.version
+git_commit_ref='54afe8d9468359e80322283cc5b6b7cd0cb6a630'
 
 usage()
 {
@@ -10,25 +12,46 @@ usage()
     echo "copy            default is to deploy scripts"
 }
 
+# search ARTLOG_TOOLBOX= within $toolboxparam file, no bash expansion is done.
+# copied from scripts/locate_artlog_toolbox.sh but required here since bootstrap
+extract_from_toolbox_param()
+{
+    local toolboxparam=$1
+    if [[ -f $toolboxparam ]]
+    then
+	echo "extract ARTLOG_TOOLBOX from $toolboxparam" >&2
+	while read LINE
+	do
+	    if [[ $LINE =~ ARTLOG_TOOLBOX=(.*) ]]
+	    then
+		ARTLOG_TOOLBOX=${BASH_REMATCH[1]}
+	    fi
+	done <$toolboxparam
+    fi
+}
+
 function update_migration()
 {
     if [[ ! -e deploy.version ]]
     then
-	# .git/refs/heads/master >deploy.version
-	echo "54afe8d9468359e80322283cc5b6b7cd0cb6a630" >deploy.version
+	echo "$git_commit_ref" >deploy.version
     fi
 
     deployed_version=$(< deploy.version)
 
-    if [[ "$deployed_version" == "54afe8d9468359e80322283cc5b6b7cd0cb6a630" ]]
+    if [[ "$deployed_version" == "$git_commit_ref" ]]
     then
 	if [[ ! -e README.md ]]
 	then
-	    echo "[INFO] migration README -> README.md"
-	    mv README README.md
+            if [[ -e README ]]
+            then
+	        echo "[INFO] migration README -> README.md"
+	        mv README README.md
+            fi
 	fi
     fi
 
+    # really update deploy.version
     cp ${A_TOOLBOX}/.git/refs/heads/master deploy.version
 }
 
@@ -39,7 +62,7 @@ then
     echo "[INFO] use -help to get help"
     exit 0
 fi
-    
+
 while [[ $# -gt 0 ]]
 do
     case $1 in
@@ -63,6 +86,17 @@ done
 
 
 PROJECT_DIR=$(pwd)
+
+# ARTLOG_TOOLBOX can be exported as environment variable
+if [[ -z $ARTLOG_TOOLBOX ]]
+then
+    # or can be set within local project toolbox.param
+    if [[ -e toolbox.param ]]
+    then
+        echo "[WARNING] This is not the first deploy since toolbox.param file exists here" >&2
+        extract_from_toolbox_param toolbox.param
+    fi
+fi
 
 if [[ -z $ARTLOG_TOOLBOX ]]
 then
@@ -122,4 +156,3 @@ then
     echo "Update migration"
     update_migration
 fi
-
